Implement support for VAT in refunds
The actual VAT calculation will have to be done manually when
a partial refund is done, since we can mix VAT amounts. This
is unlikely to happen though, and refunds in general are quite
uncommon, so we'll live with it.

In passing, implement a nicer looking refund notice since we
need to have VAT on it, including adding to the refund
notice that the payment was returned to the original form of
payment, as well as a hint to what that method was (including
the type of credit card when available).

NOTE! This makes an incompatible change to the PDF refund API,
the same way that 5c2989d did for invoices. If VAT is not used,
nothing other than just accepting the parameters and ignoring
them is required.

+++ postgresqleu/adyen/util.py
@@ -369,7 +369,7 @@ class AdyenAPI(object):
 		apiparam = {
 			'merchantAccount': settings.ADYEN_MERCHANTACCOUNT,
 			'modificationAmount': {
-				'value': amount * 100, # "minor units", so cents!
+				'value': int(amount * 100), # "minor units", so cents!
 				'currency': settings.CURRENCY_ISO,
 			},
 			'originalReference': transreference,
